refactor: Add test directory for learning tests
